% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_init.R
\name{con_init}
\alias{con_init}
\title{Initialize DataBase connections}
\usage{
con_init(
  creds = "creds",
  db_name = NULL,
  db_host = NULL,
  db_user = NULL,
  db_pass = NULL
)
}
\arguments{
\item{creds}{a filepath to the credentials file}

\item{db_name}{name of the dendrogreif database}

\item{db_host}{a character string address of the dendrogreif database server}

\item{db_user}{user name registered on the dendrogreif database server}

\item{db_pass}{password of the db_user}
}
\value{
An active db connection which needs to be used for other activities here.
}
\description{
Initialize DataBase connections
}
\details{
If you have credentials file from the system admin you can use it to initialize the database connection. Otherwise parameters to access DendroGreif system must be named manually.
}
\examples{
dg_con <- con_init()

}
