% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from dendrogreif database based from location_id and variable_id}
\usage{
get_data(
  con = NULL,
  location_ids = NULL,
  variable_ids = NULL,
  minutes = 0:59,
  source = "obs_30",
  start = "2013-01-01",
  end = Sys.Date(),
  toclean = "raw"
)
}
\arguments{
\item{con}{initialized connection}

\item{location_ids}{location_ids from list_locations()}

\item{variable_ids}{variable_ids from list_variables()}

\item{minutes}{range of the minutes to retrieve. Only relevant for the raw data series.}

\item{source}{source table of the data}

\item{start}{iso8601 timestamp of the data beginning}

\item{end}{iso8601 timestamp of the data end}

\item{toclean}{"clean", "raw" or "both" data}
}
\value{
A data.frame with monitoring locations, labels, descriptions and available variables. Location_id and variable_id combinations can be used to get_data() from the database.
}
\description{
Get data from dendrogreif database based from location_id and variable_id
}
\details{
This is the main function of the package, which requests the data from the database system. Please take a look at the examples below for common usage. Minutes parameter is only relevant for the raw SapFlow series. Otherwise default value is fine.
}
\examples{
## initialize database connection from credentials file.
conn <- con_init("creds")

#list all sites in the database
list_sites(conn)

demmin_locations <- list_locations(conn, 6)
demmin_locations \%>\% head(20) \%>\% knitr::kable()
#get single tree dendrometer
data_ind <- get_data(conn, location_ids = 67, variable_ids = 1)
print(head(data_ind))

#get all dendrometer series from the single site
data_all_dend <- get_data(conn, location_ids = get_locations(conn, site_ids=6), variable_ids = 1)
print(head(data_all_dend))

#get all series from the single site
data_all <- get_data(conn, location_ids = get_locations(conn, site_ids=6), variable_ids = 1:200)
print(head(data_all))

}
