% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{Get all locations from the site in the vector for further use}
\usage{
get_locations(...)
}
\arguments{
\item{...}{arguments passed through to the list_locations, including active connection.}
}
\value{
A vector with monitoring location_ids
}
\description{
Get all locations from the site in the vector for further use
}
\details{
The function returns the vectr of the list location_ids. Can be used to select entire dataset from the site(s).
}
\examples{
dg_con <- con_init()
sites <- list_sites(dg_con)
res <- get_data(dg_con, get_locations(con=dg_con, site_ids = 6), 1)

}
