% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables}
\alias{get_variables}
\title{Get all variables observed at the given locations for the further use}
\usage{
get_variables(...)
}
\arguments{
\item{...}{arguments passed through to the list_variables, including active connection.}
}
\value{
A vector with variable_ids
}
\description{
Get all variables observed at the given locations for the further use
}
\details{
The function returns the vectorized variables_ids from the list_variables functions. Can be used to select entire dataset from the site(s).
}
\examples{
dg_con <- con_init()
sites <- list_sites(dg_con)
loc_ids <- get_locations(con=dg_con, site_ids = 6)
var_ids <- get_variables(con=dg_con, location_ids = loc_ids)
res <- get_data(dg_con, loc_ids, var_ids)

}
