% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_locations.R
\name{list_locations}
\alias{list_locations}
\title{List all locations for the sites}
\usage{
list_locations(con = NULL, site_ids)
}
\arguments{
\item{con}{an active connection to the DB initialized via con_init().}

\item{site_ids}{site_ids from list_sites()}
}
\value{
A tibble with monitoring locations, labels, descriptions and available variables. Location_id and variable_id combinations can be used to get_data() from the database.
}
\description{
List all locations for the sites
}
\examples{
dg_con <- con_init()
sites <- list_sites(dg_con)
locations <- list_locations(dg_con, c(2,4))
}
