% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sites.R
\name{list_sites}
\alias{list_sites}
\title{List all sites in the database}
\usage{
list_sites(con = NULL, ...)
}
\arguments{
\item{con}{connection to the PostGreSQL DendroGreif database server}
}
\value{
A data.frame with complete 'sites' data table from the database
}
\description{
List all sites in the database
}
\details{
If the con is NULL the function will try to initialize the connection to the dendrogreif database.
}
\examples{
dg_con <- con_init()
list_sites(dg_con)

}
