% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_variables.R
\name{list_variables}
\alias{list_variables}
\title{List all variables for the given locations}
\usage{
list_variables(con = NULL, location_ids)
}
\arguments{
\item{location_ids}{location_ids from list_locations()}
}
\value{
A tibble with monitoring locations, labels, descriptions and available variables. Location_id and variable_id combinations can be used to get_data() from the database.
}
\description{
List all variables for the given locations
}
\details{
The function looks for the available variable_ids (for example Dendrometers, Temperature etc...) for the locations and returns the database table entry for them.
}
\examples{
dg_con <- con_init()
sites <- list_sites(dg_con)
locations <- list_locations(dg_con, c(2,4))
variables <- list_variables(dg_con, c(2,5,6))
}
